requirements_windows.txt
fastapi>=0.100.0
uvicorn>=0.20.0
numpy>=1.21.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-magic-bin>=0.4.14
requests>=2.25.0
# psycopg2-binary>=2.9.0  # Only needed for PostgreSQL
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0 